Abstracted away search methods to a Search service.
Added a new BlogSearchService class to service and registered in Startup.cs
Removed search method from Post controller and linked it to the search class.
Tested it and it worked fine.
