+++ Controllers/PostsController.cs
@@ -21,13 +21,15 @@ namespace FlynnNotesBlog.Controllers
         private readonly ISlugService _slugService;
         private readonly IImageService _imageService;
         private readonly UserManager<BlogUser> _userManager;
+        private readonly BlogSearchService _blogSearchService;
 
-        public PostsController(ApplicationDbContext context, ISlugService slugService, IImageService imageService, UserManager<BlogUser> userManager)
+        public PostsController(ApplicationDbContext context, ISlugService slugService, IImageService imageService, UserManager<BlogUser> userManager, BlogSearchService blogSearchService)
         {
             _context = context;
             _slugService = slugService;
             _imageService = imageService;
             _userManager = userManager;
+            _blogSearchService = blogSearchService;
         }
 
         public async Task<IActionResult> SearchIndex(int? page, string searchTerm)
@@ -37,22 +39,8 @@ namespace FlynnNotesBlog.Controllers
             var pageNumber = page ?? 1;
             var pageSize = 5;
 
-            var posts = _context.Posts.Where(p => p.ReadyStatus == ReadyStatus.ProductionReady).AsQueryable();
-            if(searchTerm != null)
-            {
-                searchTerm = searchTerm.ToLower();
-                posts = posts.Where(
-                    p => p.Title.ToLower().Contains(searchTerm) ||
-                    p.Abstract.ToLower().Contains(searchTerm) ||
-                    p.Content.ToLower().Contains(searchTerm) ||
-                    p.Comments.Any(c => c.Body.ToLower().Contains(searchTerm) ||
-                                        c.ModeratedBody.ToLower().Contains(searchTerm) ||
-                                        c.Author.FirstName.ToLower().Contains(searchTerm) ||
-                                        c.Author.LastName.ToLower().Contains(searchTerm) ||
-                                        c.Author.Email.ToLower().Contains(searchTerm)));
-            }
+            var posts = _blogSearchService.Search(searchTerm);
 
-            posts = posts.OrderByDescending(p => p.Created);
             return View(await posts.ToPagedListAsync(pageNumber, pageSize));
 
         }
